requirements.txt
python-pptx
beautifulsoup4
scrapy
pyyaml
loginform
selenium